Switch database from MySql to PostgraseSql

+++ src/main/java/com/wn/loanapp/dto/RoleDTO.java
@@ -2,23 +2,23 @@ package com.wn.loanapp.dto;
 
 public class RoleDTO {
 
-	private int id;
+	private long id;
 	
 	private String role;
 	
 	private String roleName;
-
+	
 	/**
 	 * @return the id
 	 */
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
 	/**
 	 * @param id the id to set
 	 */
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
